requirements.txt
streamlit 
requests
json
pandas 
numpy 
matplotlib.pyplot 
fastapi 
pydantic
jobliuvicorn